Move `is_forward_transaction` away from `codec.rs`
