+++ chain/network/src/peer/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod codec;
 pub(crate) mod peer_actor;
 mod rate_counter;
 mod tracker;
+pub(crate) mod utils;
