+++ chain/network/src/peer/peer_actor.rs
@@ -1,6 +1,7 @@
 use crate::common::message_wrapper::ActixMessageWrapper;
-use crate::peer::codec::{self, Codec};
+use crate::peer::codec::Codec;
 use crate::peer::tracker::Tracker;
+use crate::peer::utils;
 use crate::routing::edge::{Edge, PartialEdgeInfo};
 use crate::stats::metrics::{self, NetworkMetrics};
 use crate::types::{
@@ -557,7 +558,7 @@ impl PeerActor {
     /// Check whenever we exceeded number of transactions we got since last block.
     /// If so, drop the transaction.
     fn should_we_drop_msg_without_decoding(&self, msg: &Vec<u8>) -> bool {
-        if codec::is_forward_transaction(msg).unwrap_or(false) {
+        if utils::is_forward_transaction(msg).unwrap_or(false) {
             let r = self.txns_since_last_block.load(Ordering::Acquire);
             if r > MAX_TRANSACTIONS_PER_BLOCK_MESSAGE {
                 return true;
